+++ src/Tasks/TaskSelectHyperplanePointsECP.cpp
@@ -57,7 +57,7 @@ void TaskSelectHyperplanePointsECP::run(std::vector<SolutionPoint> solPoints)
     std::vector<std::tuple<int, NumericConstraintValue>> selectedNumericValues;
     std::vector<std::tuple<int, NumericConstraintValue>> nonconvexSelectedNumericValues;
 
-    for(int i = 0; i < solPoints.size(); i++)
+    for(size_t i = 0; i < solPoints.size(); i++)
     {
         auto numericConstraintValues = env->reformulatedProblem->getFractionOfDeviatingNonlinearConstraints(
             solPoints.at(i).point, 0.0, constraintSelectionFactor);
